refactor: move Unit and Region to jvm.classes
`Unit` and `Region` become `GenUnit` and `GenRegion`, each owning its descriptor
instead of a case in the central `desc` match. `BackendObjType.scala` drops from
838 to 642 lines.

This closes the cycle left open in #13131: `GenUncaughtExceptionHandler` and
`GenRegion` now reference each other across the `jvm.classes` package rather than
across the package boundary.

`toDescriptor` came from the `BackendObjType` trait, so the two `GenUnit` uses of
it in `GenExpression` become `GenUnit.desc.descriptorString()`.
